chore(api): split test fakes per aggregate (868kvw2hr)
Fallow I. fakes.ts was the repo's top churn hotspot — 396 LOC, 7 dependents,
417 lines churned over 7 commits — so every schema change rippled through all
five API suites at once.

The aggregates now own their own files (recipeFakes, ingredientFakes,
supplierFakes) over a shared state module; fakes.ts stays the entry point and
re-exports them, so no test file changes. Search spans two aggregates and
stays in the entry point.

Folds the clones the finding named:
- the findIndex/NotFoundError pair in every update and delete becomes
  requireOwnedIndex()
- the recipe detail row built in both seedRecipe and create becomes
  toRecipeDetail()
- the suppliers-stripping display mapping used by ingredients and search
  becomes toIngredientDisplay()

Verified: pnpm --filter api test green, 80 tests, no assertion touched;
pnpm build and pnpm lint green; fallow dupes reports no clone group in fakes.
